requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-debug-toolbar==5.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
whitenoise==6.9.0
